requirements.txt
# Core dependencies
clinlp>=0.9.4
spacy>=3.7.0
transformers>=4.35.0
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Dutch NLP models and resources
https://huggingface.co/DTAI-KULeuven/robbert-2023-dutch-large/resolve/main/pytorch_model.bin

# API and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database and storage
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
redis>=5.0.0

# Document processing
python-docx>=0.8.11
PyPDF2>=3.0.0
python-magic>=0.4.27
openpyxl>=3.1.0

# Security and encryption
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Monitoring and logging
structlog>=23.0.0
prometheus-client>=0.18.0
sentry-sdk>=1.38.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.6.0
pre-commit>=3.5.0

# Deployment
gunicorn>=21.2.0
docker>=6.1.0
kubernetes>=28.1.0

# Data science and analysis
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.66.0
click>=8.1.0
rich>=13.6.0
typer>=0.9.0

